updating ManglerDBAdapter to prep for multiple tables
git-svn-id: http://svn.mangler.org/mangler/trunk@991 f654edcb-a5fb-4f28-a0e7-68c327795dae

+++ android/src/org/mangler/android/ManglerDBAdapter.java
@@ -34,13 +34,21 @@ import android.util.Log;
 
 public class ManglerDBAdapter {
 
-	public static final String KEY_ROWID = "_id";
-    public static final String KEY_SERVERNAME = "servername";
-    public static final String KEY_HOSTNAME = "hostname";
-    public static final String KEY_PORTNUMBER = "portnumber";
-    public static final String KEY_PASSWORD = "password";
-    public static final String KEY_USERNAME = "username";
-    public static final String KEY_PHONETIC = "phonetic";
+	public static final String KEY_SERVERS_ROWID = "_id";
+    public static final String KEY_SERVERS_SERVERNAME = "servername";
+    public static final String KEY_SERVERS_HOSTNAME = "hostname";
+    public static final String KEY_SERVERS_PORTNUMBER = "portnumber";
+    public static final String KEY_SERVERS_PASSWORD = "password";
+    public static final String KEY_SERVERS_USERNAME = "username";
+    public static final String KEY_SERVERS_PHONETIC = "phonetic";
+    
+    public static final String KEY_VOLUME_SERVERID = "_server_id";
+    public static final String KEY_VOLUME_USERNAME = "username";
+    public static final String KEY_VOLUME_LEVEL = "level";
+
+    public static final String KEY_PASSWORD_SERVERID = "_server_id";
+    public static final String KEY_PASSWORD_CHANNEL = "channel";
+    public static final String KEY_PASSWORD_PASSWORD = "password";
 
     private static final String TAG = "ManglerDBAdapter";
     
@@ -49,11 +57,17 @@ public class ManglerDBAdapter {
     
     // SQL string for creating database
     private static final String DATABASE_CREATE =
-        "create table servers (_id integer primary key autoincrement, servername text not null, hostname text not null, portnumber integer not null, password text not null, username text not null, phonetic text not null);";
+        "create table servers (_id integer primary key autoincrement, servername text not null, hostname text not null, portnumber integer not null, password text not null, username text not null, phonetic text not null);" +
+        "create table volume (_server_id integer, username text not null, level integer not null);" +
+    	"create table password (_server_id integer, channel text not null, channel text not null);";
 
     private static final String DATABASE_NAME = "manglerdata";
     private static final String DATABASE_SERVER_TABLE = "servers";
-    private static final int DATABASE_VERSION = 2;
+    private static final String DATABASE_VOLUME_TABLE = "volume";
+    private static final String DATABASE_PASSWORD_TABLE = "password";
+
+
+    private static final int DATABASE_VERSION = 3;
 
     private final Context context;
 
@@ -73,8 +87,14 @@ public class ManglerDBAdapter {
         public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
             Log.w(TAG, "Upgrading database from version " + oldVersion + " to "
                     + newVersion + ", which will destroy all old data");
-            db.execSQL("DROP TABLE IF EXISTS servers");
-            onCreate(db);
+            if (oldVersion < 2) {
+            	db.execSQL("DROP TABLE IF EXISTS servers");
+            	onCreate(db);
+            }
+            if (oldVersion < 3) {
+            	db.execSQL("create table volume (_server_id integer, username text not null, level integer not null);");
+        		db.execSQL("create table password (_server_id integer, channel text not null, channel text not null);");
+            }
         }
     }
 
@@ -121,12 +141,12 @@ public class ManglerDBAdapter {
      */
     public long createServer(String servername, String hostname, int portnumber, String password, String username, String phonetic) {
         ContentValues initialValues = new ContentValues();
-        initialValues.put(KEY_SERVERNAME, servername);
-        initialValues.put(KEY_HOSTNAME, hostname);
-        initialValues.put(KEY_PORTNUMBER, portnumber);
-        initialValues.put(KEY_PASSWORD, password);
-        initialValues.put(KEY_USERNAME, username);
-        initialValues.put(KEY_PHONETIC, phonetic);
+        initialValues.put(KEY_SERVERS_SERVERNAME, servername);
+        initialValues.put(KEY_SERVERS_HOSTNAME, hostname);
+        initialValues.put(KEY_SERVERS_PORTNUMBER, portnumber);
+        initialValues.put(KEY_SERVERS_PASSWORD, password);
+        initialValues.put(KEY_SERVERS_USERNAME, username);
+        initialValues.put(KEY_SERVERS_PHONETIC, phonetic);
 
         return db.insert(DATABASE_SERVER_TABLE, null, initialValues);
     }
@@ -141,12 +161,12 @@ public class ManglerDBAdapter {
     public long cloneServer(long rowId) {
     	Cursor cursor = fetchServer(rowId);
         ContentValues initialValues = new ContentValues();
-        initialValues.put(KEY_SERVERNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERNAME)));
-        initialValues.put(KEY_HOSTNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_HOSTNAME)));
-        initialValues.put(KEY_PORTNUMBER, cursor.getInt(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_PORTNUMBER)));
-        initialValues.put(KEY_PASSWORD, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_PASSWORD)));
-        initialValues.put(KEY_USERNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_USERNAME)));
-        initialValues.put(KEY_PHONETIC, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_PHONETIC)));
+        initialValues.put(KEY_SERVERS_SERVERNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_SERVERNAME)));
+        initialValues.put(KEY_SERVERS_HOSTNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_HOSTNAME)));
+        initialValues.put(KEY_SERVERS_PORTNUMBER, cursor.getInt(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_PORTNUMBER)));
+        initialValues.put(KEY_SERVERS_PASSWORD, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_PASSWORD)));
+        initialValues.put(KEY_SERVERS_USERNAME, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_USERNAME)));
+        initialValues.put(KEY_SERVERS_PHONETIC, cursor.getString(cursor.getColumnIndexOrThrow(ManglerDBAdapter.KEY_SERVERS_PHONETIC)));
 
         return db.insert(DATABASE_SERVER_TABLE, null, initialValues);
     }
@@ -159,7 +179,7 @@ public class ManglerDBAdapter {
      */
     public boolean deleteServer(long rowId) {
 
-        return db.delete(DATABASE_SERVER_TABLE, KEY_ROWID + "=" + rowId, null) > 0;
+        return db.delete(DATABASE_SERVER_TABLE, KEY_SERVERS_ROWID + "=" + rowId, null) > 0;
     }
     
     /**
@@ -169,8 +189,8 @@ public class ManglerDBAdapter {
      */
     public Cursor fetchServers() {
 
-        return db.query(DATABASE_SERVER_TABLE, new String[] {KEY_ROWID, KEY_SERVERNAME, 
-        		KEY_HOSTNAME, KEY_PORTNUMBER, KEY_PASSWORD, KEY_USERNAME, KEY_PHONETIC}, null, null, null, null, null);
+        return db.query(DATABASE_SERVER_TABLE, new String[] {KEY_SERVERS_ROWID, KEY_SERVERS_SERVERNAME, 
+        		KEY_SERVERS_HOSTNAME, KEY_SERVERS_PORTNUMBER, KEY_SERVERS_PASSWORD, KEY_SERVERS_USERNAME, KEY_SERVERS_PHONETIC}, null, null, null, null, null);
     }
     
     /**
@@ -183,9 +203,9 @@ public class ManglerDBAdapter {
     public Cursor fetchServer(long rowId) throws SQLException {
 
         Cursor cursor =
-                db.query(true, DATABASE_SERVER_TABLE, new String[] {KEY_ROWID,
-                        KEY_SERVERNAME, KEY_HOSTNAME, KEY_PORTNUMBER, KEY_PASSWORD, KEY_USERNAME, KEY_PHONETIC}, 
-                        KEY_ROWID + "=" + rowId, null, null, null, null, null);
+                db.query(true, DATABASE_SERVER_TABLE, new String[] {KEY_SERVERS_ROWID,
+                        KEY_SERVERS_SERVERNAME, KEY_SERVERS_HOSTNAME, KEY_SERVERS_PORTNUMBER, KEY_SERVERS_PASSWORD, KEY_SERVERS_USERNAME, KEY_SERVERS_PHONETIC}, 
+                        KEY_SERVERS_ROWID + "=" + rowId, null, null, null, null, null);
         if (cursor != null) {
             cursor.moveToFirst();
         }
@@ -207,13 +227,13 @@ public class ManglerDBAdapter {
      */
     public boolean updateServer(long rowId, String servername, String hostname, int portnumber, String password, String username, String phonetic) {
         ContentValues args = new ContentValues();
-        args.put(KEY_SERVERNAME, servername);
-        args.put(KEY_HOSTNAME, hostname);
-        args.put(KEY_PORTNUMBER, portnumber);
-        args.put(KEY_PASSWORD, password);
-        args.put(KEY_USERNAME, username);
-        args.put(KEY_PHONETIC, phonetic);
+        args.put(KEY_SERVERS_SERVERNAME, servername);
+        args.put(KEY_SERVERS_HOSTNAME, hostname);
+        args.put(KEY_SERVERS_PORTNUMBER, portnumber);
+        args.put(KEY_SERVERS_PASSWORD, password);
+        args.put(KEY_SERVERS_USERNAME, username);
+        args.put(KEY_SERVERS_PHONETIC, phonetic);
 
-        return db.update(DATABASE_SERVER_TABLE, args, KEY_ROWID + "=" + rowId, null) > 0;
+        return db.update(DATABASE_SERVER_TABLE, args, KEY_SERVERS_ROWID + "=" + rowId, null) > 0;
     }
 }
